feat(#166): add pyproject.toml to shared packages and restructure creator-provider
- Add pyproject.toml for creator-domain, creator-service, creator-provider
- Move creator-provider source into creator_provider/ subdirectory for
  standard Python packaging (fixes editable install MAPPING issue)
- Update test imports to use proper package paths instead of sys.path hacks
- Add *.egg-info/ to .gitignore, remove obsolete symlink entries
- All 12 provider tests + 8 domain tests + 129 service tests pass
