+++ .gitignore
@@ -1,4 +1,5 @@
 # Python
+*.egg-info/
 __pycache__/
 *.py[cod]
 *.pyo
@@ -32,8 +33,3 @@ Thumbs.db
 # Artifacts
 data/artifacts/**
 !data/artifacts/.gitkeep
-packages/creator-domain/creator-domain
-packages/creator-service/creator-service
-packages/creator_domain
-packages/creator_provider
-packages/creator_service
