+++ packages/creator-provider/creator_provider/stt/external_stt_provider.py
