+++ packages/creator-provider/creator_provider/stt/whisper_provider.py
